fix: enterTol() function changed from float type to double type. Now, calculateNaturalExponential() function 2nd parameter is double type.

+++ schoolwork-part1/naturalExponential.h
@@ -3,9 +3,9 @@
 
 #include "factorial.h"
 
-double calculateNaturalExponential(int x, float tol);
+double calculateNaturalExponential(int x, double tol);
 int enterNumber();
-float enterTol();
+double enterTol();
 int raiseNumber(int number, int exponent);
 
 #endif // NATURALEXPONENTIAL
